Working factories using faker for all models (no relationships)

+++ database/migrations/2020_11_06_211536_create_posts_table.php
@@ -21,7 +21,7 @@ class CreatePostsTable extends Migration
             $table->string('description');
             $table->boolean('solved'); // Whether an accepted answer has been found
 
-            $table->unsignedInteger('author_id'); // User the post belongs to
+            $table->unsignedBigInteger('author_id'); // User the post belongs to
 
         });
     }
